requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
azure-cognitiveservices-vision-computervision
msrest
openai
python-multipart
